moved gaussian kernel functions to utils.py
